requirements.txt
load_dotenv
databases
asyncpg
aiodns
aiohttp[speedups]
httpx
trio
pytest
pandas
openpyxl
xlrd
requests
aiohttp
plotly
dash
dash_bootstrap_components
asgi_lifespan
tortoise-orm[asyncpg]
uvicorn
fastapi
fastapi-crudrouter
pysha
web3
coinaddrvalidator
bitcoinaddress
cryptos
bitcoin
alembic

